+++ lib/overcommit/plugins/pre_commit/css_linter.rb
@@ -11,7 +11,6 @@ module Overcommit::GitHook
       paths = staged.map { |s| s.path }.join(' ')
 
       output = `rhino #{CSS_LINTER_PATH} --quiet --format=compact #{paths} | grep 'Error - '`
-      staged.each { |s| output = s.filter_string(output) }
       return (output !~ /Error - (?!Unknown @ rule)/ ? :good : :bad), output
     end
   end